renamed to end2 and created a resource_profile module
