+++ README.md
@@ -104,6 +104,7 @@ if __name__ == '__main__':
 ## TODO:
 - [] change suites to be file path instead of dot notation
 - [] support async fixtures
+- [] make runner use suitelogmanager again
 - [x] .testingrc or maybe setting.conf (have this file as a profile with setting about how to configure runner)
     - [x] max threads
     - [x] max sub folder logs
