+++ end2/exceptions.py
